[ANDROSDK-1964] Move LegendSetFields to network package
