+++ core/src/main/java/org/hisp/dhis/android/core/indicator/internal/IndicatorFields.kt
@@ -35,7 +35,7 @@ import org.hisp.dhis.android.core.common.objectstyle.internal.ObjectStyleFields
 import org.hisp.dhis.android.core.indicator.Indicator
 import org.hisp.dhis.android.core.indicator.IndicatorTableInfo.Columns
 import org.hisp.dhis.android.core.legendset.LegendSet
-import org.hisp.dhis.android.core.legendset.internal.LegendSetFields
+import org.hisp.dhis.android.network.legendset.LegendSetFields
 
 internal object IndicatorFields : BaseFields<Indicator>() {
     const val LEGEND_SETS = "legendSets"
