+++ core/src/main/java/org/hisp/dhis/android/core/legendset/LegendSetCollectionRepository.kt
@@ -34,7 +34,7 @@ import org.hisp.dhis.android.core.arch.repositories.filters.internal.FilterConne
 import org.hisp.dhis.android.core.arch.repositories.filters.internal.StringFilterConnector
 import org.hisp.dhis.android.core.arch.repositories.scope.RepositoryScope
 import org.hisp.dhis.android.core.legendset.internal.LegendChildrenAppender
-import org.hisp.dhis.android.core.legendset.internal.LegendSetFields
+import org.hisp.dhis.android.network.legendset.LegendSetFields
 import org.hisp.dhis.android.core.legendset.internal.LegendSetStore
 import org.koin.core.annotation.Singleton
